+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Added support for executing Basilisp namespaces directly via `basilisp run` and by `python -m` (#791)
  * Added the `memoize` core fn (#812)
  * Added support for `thrown-with-msg?` assertions to `basilisp.test/is` (#831)
+ * Added support for reading scientific notation literals, octal and hex integer literals, and arbitrary base (2-36) integer literals (#769)
 
 ### Changed
  * Optimize calls to Python's `operator` module into their corresponding native operators (#754)